backend/requirements.txt
cchardet
aiodns
aiohttp
PyYAML
